backend/requirements.txt
django
djangorestframework
daphne
wheel
psycopg2-binary
django-cors-headers
channels
redis
Pillow
channels_redis